github fr - 複製/requirements.txt
openai
pyyaml
sounddevice
soundfile
numpy
requests
whisper-openai
